% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash.R
\name{adjust_by_t}
\alias{adjust_by_t}
\title{Returns adjusted sebetahat's based on t likelihood so that we can use a normal likelihood.}
\usage{
adjust_by_t(betahat, sebetahat, df)
}
\arguments{
\item{betahat}{The estimates of the effects.}

\item{sebetahat}{The estimates of the standard errors of \code{betahat}.}

\item{df}{The degrees of freedom of the t. Can either be of length 1 or the same length of
\code{betahat}.}
}
\description{
Returns adjusted sebetahat's based on t likelihood so that we can use a normal likelihood.
}
\author{
David Gerard
}
