% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwash.R
\name{back_fix}
\alias{back_fix}
\title{Fixed point iteration for BACKWASH.}
\usage{
back_fix(par_vec, betahat_ols, S_diag, Amat, tau2_seq, lambda_seq)
}
\arguments{
\item{par_vec}{A huge vector of parameters whose elements are in
the following order: pivec, mubeta_matrix, sig2beta_matrix,
gamma_mat, muv, Sigma_v, phi, xi}

\item{betahat_ols}{A vector of numerics. The oridinary least
squares estimates of the regression coefficients.}

\item{S_diag}{A vector of positive numerics. The standard errors of
\code{betahat_ols}.}

\item{Amat}{The A matrix for the variational EM.}

\item{tau2_seq}{The known grid of prior mixing variances.}

\item{lambda_seq}{A vector of numerics greater than 1. The penalties for the prior mixing proportions.}
}
\description{
This is mostly so that I can use the SQUAREM package.
}
\author{
David Gerard
}

