% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwash.R
\name{back_update_phi}
\alias{back_update_phi}
\title{Update for the "g" hyperparameter.}
\usage{
back_update_phi(betahat_ols, S_diag, Amat, mubeta, muv, Sigma_v)
}
\arguments{
\item{betahat_ols}{A vector of numerics. The ordinary least
squares estimates of the regression coefficients.}

\item{S_diag}{A vector of positive numerics. The standard errors of
\code{betahat_ols}.}

\item{Amat}{The A matrix for the variational EM.}

\item{mubeta}{The current means of the betas.}

\item{muv}{A matrix of one column of numerics. The current mean of
v.}

\item{Sigma_v}{The current covarianc matrix of the latent v.}
}
\description{
Update for the "g" hyperparameter.
}
\author{
David Gerard
}
