% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwash.R
\name{back_update_pi}
\alias{back_update_pi}
\title{Update for the prior mixing proportions.}
\usage{
back_update_pi(gamma_mat, lambda_seq)
}
\arguments{
\item{gamma_mat}{The current values of the variational mixing
proportions. The number of columns is the number of
observations and the number of rows is the number of mixture
components. The rows must sum to one.}

\item{lambda_seq}{A vector of numerics greater than 1. The
penalties on the pi's.}
}
\description{
Update for the prior mixing proportions.
}
\author{
David Gerard
}
