% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwash.R
\name{back_update_qbeta}
\alias{back_update_qbeta}
\title{Update for the variational density of beta}
\usage{
back_update_qbeta(betahat_ols, S_diag, Amat, pivec, tau2_seq, muv, xi, phi)
}
\arguments{
\item{betahat_ols}{A vector of numerics. The oridinary least
squares estimates of the regression coefficients.}

\item{S_diag}{A vector of positive numerics. The standard errors of
\code{betahat_ols}.}

\item{Amat}{The A matrix for the variational EM.}

\item{pivec}{The current values of the mixing proportions.}

\item{tau2_seq}{The known grid of prior mixing variances.}

\item{muv}{A matrix of one column of numerics. The current mean of
v.}

\item{xi}{A positive numeric. The current value of xi.}

\item{phi}{A numeric. The "g" hyperparameter.}
}
\description{
Update for the variational density of beta
}
\author{
David Gerard
}
