% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwash.R
\name{back_update_v}
\alias{back_update_v}
\title{Update for the variational density of v.}
\usage{
back_update_v(betahat_ols, S_diag, Amat, mubeta, xi, phi)
}
\arguments{
\item{betahat_ols}{A vector of numerics. The oridinary least
squares estimates of the regression coefficients.}

\item{S_diag}{A vector of positive numerics. The standard errors of
\code{betahat_ols}.}

\item{Amat}{The A matrix for the variational EM.}

\item{mubeta}{The current means of the betas.}

\item{xi}{A positive numeric. The current value of xi.}

\item{phi}{A numeric. The "g" hyperparameter.}
}
\description{
Update for the variational density of v.
}
\author{
David Gerard
}
