% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwash.R
\name{back_update_xi}
\alias{back_update_xi}
\title{Update for the variance scaling parameter.}
\usage{
back_update_xi(betahat_ols, S_diag, Amat, mubeta, mubeta_matrix,
  sig2beta_matrix, gamma_mat, muv, Sigma_v, phi)
}
\arguments{
\item{betahat_ols}{A vector of numerics. The oridinary least
squares estimates of the regression coefficients.}

\item{S_diag}{A vector of positive numerics. The standard errors of
\code{betahat_ols}.}

\item{Amat}{The A matrix for the variational EM.}

\item{mubeta}{The current means of the betas}

\item{mubeta_matrix}{The current mixing means of the variational
densities of the betas.}

\item{sig2beta_matrix}{The current mixing variances of the
variational densities of the betas.}

\item{gamma_mat}{The current mixing proportions of the variational
densities of the betas.}

\item{muv}{A matrix of one column of numerics. The current mean of
v.}

\item{Sigma_v}{The current covariance matrix of the latent v.}

\item{phi}{The current "g" hyperparameter.}
}
\description{
Update for the variance scaling parameter.
}
