% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwash.R
\name{backwash_second_step}
\alias{backwash_second_step}
\title{Second step of the backwash procedure.}
\usage{
backwash_second_step(
  betahat_ols,
  S_diag,
  alpha_tilde,
  tau2_seq,
  lambda_seq,
  pi_init_type = c("zero_conc", "uniform", "random"),
  scale_var = TRUE,
  sprop = 0,
  var_inflate_pen = 0,
  verbose = TRUE
)
}
\arguments{
\item{betahat_ols}{A vector of numerics. The ordinary least
squares estimates of the regression coefficients.}

\item{S_diag}{A vector of positive numerics. The standard errors of
\code{betahat_ols}.}

\item{alpha_tilde}{A matrix of numerics. The estimated coefficients
of the confounders.}

\item{tau2_seq}{A vector of positive numerics. The known grid of
prior mixing variances.}

\item{lambda_seq}{A vector of penalties for the estimate of the
mixing proportions.}

\item{pi_init_type}{How should we initialize the mixing
proportions? By concentrating on zero (\code{"zero_conc"}), by
equal weights on all mixing distributions (\code{"uniform"}),
or by sampling uniformly on the simplex (\code{"random"})?}

\item{scale_var}{A logical. Should we estimate a variance inflation
parameter (\code{TRUE}) or not (\code{FALSE})?}

\item{sprop}{If \eqn{b} is an effect and \eqn{s} is an estimated
standard error, then we model \eqn{b/s^{sprop}} as
exchangeable. The default is 0. When \code{sprop = 1}, for
identifiability reasons it must be the case that
\code{scale_var = FALSE}.}

\item{var_inflate_pen}{The penalty to apply on the variance inflation parameter.
Defaults to 0, but should be something non-zero when \code{alpha = 1}
and \code{scale_var = TRUE}.}

\item{verbose}{If \code{verbose = TRUE}, print progress of the algorithm
to the console.}
}
\description{
Second step of the backwash procedure.
}
\author{
David Gerard
}
