% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{bfa_gs_linked}
\alias{bfa_gs_linked}
\title{Simple Bayesian low rank matrix decomposition.}
\usage{
bfa_gs_linked(Y21, Y31, Y32, k, nsamp = 10000, burnin = round(nsamp/4),
  thin = 10, display_progress = TRUE, rho_0 = 0.1, alpha_0 = 0.1,
  beta_0 = 1, eta_0 = 1, tau_0 = 1, use_code = c("r", "cpp"))
}
\arguments{
\item{Y21}{Top left of matrix.}

\item{Y31}{Bottom left of matrix.}

\item{Y32}{Top right of matrix.}

\item{k}{The rank of the mean matrix.}

\item{nsamp}{A positive integer. The number of samples to draw.}

\item{burnin}{A positive integer. The number of early samples to
discard.}

\item{thin}{A positive integer. We will same the updates of
\code{Y22} every \code{keep} iteration of the Gibbs sampler.}

\item{display_progress}{A logical. Should we print a text progress
bar to keep track of the Gibbs sampler (\code{TRUE}) or not
(\code{FALSE})? Also, if \code{TRUE}, then you can interupt the
C++ code every 1\% of runtime.}

\item{rho_0}{A scalar. The prior "sample size" for the precisions.}

\item{alpha_0}{A scalar. The prior "sample size" for the mean of
the precisions.}

\item{beta_0}{A scalar. The prior mean of the precisions.}

\item{eta_0}{A scalar. The prior "sample size" for the parameter
expansion.}

\item{tau_0}{A scalar. The prior mean for the parameter expansion.
matrix.}

\item{use_code}{A character. Should we use the C++ code
(\code{"cpp"}) or the R code (\code{"r"})?}
}
\description{
"bfl" = "Bayesian factor loading"
}
\details{
This is as simple as they come. I put normal priors on the loadings
and factors and gamma priors on the precisions. The hyperparameters
are set to provide weak prior information by default.

The main difference between this version and others is that the
factors are a prior assumed to have the same variances as the data
observations. This might be distasteful to some.

This also has parameter expansion impelemented and is written in
compiled code. To see a slower version without paramaeter epansion,
go to \code{\link{bfl}}.
}
\seealso{
\code{\link{bfa_gs_linked}}.
}
\author{
David Gerard
}
