% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{bfa_wrapper}
\alias{bfa_wrapper}
\title{Wrapper for the bfa package.}
\usage{
bfa_wrapper(Y21, Y31, Y32, k, nsamp = 10000, burnin = round(nsamp/4),
  keep = 20, print_status = 500)
}
\arguments{
\item{Y21}{Top left of matrix.}

\item{Y31}{Bottom left of matrix.}

\item{Y32}{Top right of matrix.}

\item{k}{The rank of the mean matrix.}

\item{nsamp}{A positive integer. The number of samples to draw.}

\item{burnin}{A positive integer. The number of early samples to
discard.}

\item{keep}{A positive integer. We will same the updates of
\code{Y22} every \code{keep} iteration of the Gibbs sampler.}

\item{print_status}{How often should bfa print updates?}
}
\description{
This is just a wrapper for using the \code{\link[bfa]{bfa_gauss}}
from the bfa package. Default settings are used.
}
\author{
David Gerard
}
\seealso{
\code{\link[bfa]{bfa_gauss}} for the workhorse function.
}

