% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{bfl}
\alias{bfl}
\title{Simple Bayesian low rank matrix decomposition.}
\usage{
bfl(Y21, Y31, Y32, k, nsamp = 10000, burnin = round(nsamp/4), keep = 20,
  print_update = TRUE, plot_update = FALSE)
}
\arguments{
\item{Y21}{Top left of matrix.}

\item{Y31}{Bottom left of matrix.}

\item{Y32}{Top right of matrix.}

\item{k}{The rank of the mean matrix.}

\item{nsamp}{A positive integer. The number of samples to draw.}

\item{burnin}{A positive integer. The number of early samples to
discard.}

\item{keep}{A positive integer. We will same the updates of
\code{Y22} every \code{keep} iteration of the Gibbs sampler.}

\item{print_update}{A logical. Should we print a text progress bar
to keep track of the Gibbs sampler (\code{TRUE}) or not
(\code{FALSE})?}

\item{plot_update}{A logical. Should we make some plots to keep
track of the Gibbs sampler (\code{TRUE}) or not (\code{FALSE})?}
}
\description{
"bfl" = "Bayesian factor loading"
}
\details{
This is as simple as they come. I put normal priors on the loadings
and factors and gamma priors on the precisions.
}
\author{
David Gerard
}

