% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{bsvd}
\alias{bsvd}
\title{Gibbs sampler for Bayesian SVD.}
\usage{
bsvd(Y21, Y31, Y32, k, nsamp = 10000, burnin = round(nsamp/4), keep = 20,
  print_update = TRUE, plot_update = FALSE)
}
\arguments{
\item{Y21}{Top left of matrix.}

\item{Y31}{Bottom left of matrix.}

\item{Y32}{Top right of matrix.}

\item{k}{The rank of the mean matrix.}

\item{nsamp}{A positive integer. The number of samples to draw.}

\item{burnin}{A positive integer. The number of early samples to
discard.}

\item{keep}{A positive integer. We will same the updates of
\code{Y22} every \code{keep} iteration of the Gibbs sampler.}

\item{print_update}{A logical. Should we print a text progress bar
to keep track of the Gibbs sampler (\code{TRUE}) or not
(\code{FALSE})?}

\item{plot_update}{A logical. Should we make some plots to keep
track of the Gibbs sampler (\code{TRUE}) or not (\code{FALSE})?}
}
\value{
A list with the following elements:

    \code{Y22_array} A three-dimensional array containing draws of
    Y22. \code{Y22[, , i]} contains the \eqn{i}th draw of Y22.

    \code{mu_psi_phi} A matrix with three columns. The rows are the
    draws from the posterior. The first column is for the mean of
    the singular values. The second column is for the precision of
    the singular values. The last column is for the mean of the
    variances.

    \code{neff_y22} The effective sample sizes from
    \code{Y22_array} as calculated by
    \code{\link[coda]{effectiveSize}}

    \code{neff_mu_psi_phi} The effective sample sizes from
    \code{mu_psi_phi} as calculated by
    \code{\link[coda]{effectiveSize}}
}
\description{
This is a modification of the Bayesian approach from Hoff (2007) to
allow for heteroscedastic columns. We start the missing values from
the RUV4 solution.
}
\details{
The rejection sampler in \code{\link[rstiefel]{rbmf.matrix.gibbs}}
almost always stalls, so I am removing it from exports for now.
}
\references{
Hoff, P. D. (2012). Model averaging and dimension
    selection for the singular value decomposition. Journal of the
    American Statistical Association.
}
\author{
David Gerard
}
