% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{calc_lfsr}
\alias{calc_lfsr}
\title{Empirical estimate of lfsr based on posterior samples.}
\usage{
calc_lfsr(y)
}
\arguments{
\item{y}{A vector of posterior draws.}
}
\description{
Empirical estimate of lfsr based on posterior samples.
}
\author{
David Gerard
}

