% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{calc_lfsr_g}
\alias{calc_lfsr_g}
\title{Same as \code{\link{calc_lfsr}} except with a prior specification.}
\usage{
calc_lfsr_g(y, g)
}
\arguments{
\item{y}{A vector of posterior draws.}

\item{g}{A vector of priors.}
}
\description{
Same as \code{\link{calc_lfsr}} except with a prior specification.
}
\author{
David Gerard
}

