% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{calc_mean_g}
\alias{calc_mean_g}
\title{Calculate moments of pointmass rv's.}
\usage{
calc_mean_g(y, g, r = 1)
}
\arguments{
\item{y}{A vector of posterior draws.}

\item{g}{A vector of priors.}

\item{r}{A positive numeric. The moment to calculate.}
}
\description{
Calculate moments of pointmass rv's.
}
\author{
David Gerard
}
