% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{calc_quantiles_g}
\alias{calc_quantiles_g}
\title{Calculate quantiles of pointmass rv's.}
\usage{
calc_quantiles_g(y, g, quant = 0.5)
}
\arguments{
\item{y}{A vector of posterior draws.}

\item{g}{A vector of priors.}

\item{quant}{The quantile to calculate.}
}
\description{
Calculate quantiles of pointmass rv's.
}
\author{
David Gerard
}
