% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_my_p.R
\name{clean_my_p}
\alias{clean_my_p}
\title{Use the distribution of control genes' p-values to adjust all of
the p-values.}
\usage{
clean_my_p(pvalues, ctl)
}
\arguments{
\item{pvalues}{A vector of numbers between 0 and 1. The pvalues.}

\item{ctl}{A vector of logicals of the same length of
\code{pvalues}. A \code{TRUE} indicates that a p-value is from
a control gene and so should be uniformly distributed.}
}
\description{
This function will transform all p-values by the empirical cdf of
the p-values for the control genes. This will result in the
p-values of the control genes being uniformly distributed about 0
and 1.
}
\author{
David Gerard
}

