% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_my_p.R
\name{clean_my_stats}
\alias{clean_my_stats}
\title{Quantile normalize stats to their theoretical distributions.}
\usage{
clean_my_stats(stats, ctl, qfunc, ...)
}
\arguments{
\item{stats}{A vector of numerics. The statistics.}

\item{ctl}{A vector of logicals of the same length of
\code{pvalues}. A \code{TRUE} indicates that a p-value is from
a control gene and so should be uniformly distributed.}

\item{qfunc}{A quantile function for the theoretical null
distribution of the statistics in \code{stats}.}

\item{...}{More parameters to pass to \code{qfunc}.}
}
\description{
Quantile normalize stats to their theoretical distributions.
}
\author{
David Gerard
}

