% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapp.R
\name{cleapp}
\alias{cleapp}
\title{Cleaned version of LEAPP.}
\usage{
cleapp(X, Y, H, method = "hard")
}
\arguments{
\item{X}{A matrix of numerics. The covariates of interest.}

\item{Y}{A matrix of numerics. These are the response variables
where each column has its own variance. In a gene expression
study, the rows are the individuals and the columns are the
genes.}

\item{H}{Equal to X(X'X)^{-1}X'}

\item{method}{Either "hard" or "soft"}
}
\description{
Doesn't work too well.
}

