% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv4.R
\name{cruv4_multicov}
\alias{cruv4_multicov}
\title{RUV4's second step.}
\usage{
cruv4_multicov(Y2, alpha, sig_diag, ctl, R22, degrees_freedom, gls = TRUE,
  likelihood = "t")
}
\arguments{
\item{Y2}{A matrix of numerics with p rows and q columns, where q
is the number of covariates of interest and p is the number of
genes.}

\item{alpha}{A matrix of numerics of dimension p by k, where k is
the number of hidden confounders. The estimated coefficients of
the unobserved confounders.}

\item{sig_diag}{A vector of positive numerics. The estimates of the
variances.}

\item{ctl}{A vector of logicals of length \code{ncol(Y)}. If
position i is \code{TRUE} then position i is considered a
negative control.}

\item{R22}{A matrix of numerics numeric. This is the submatrix
of the R in the QR decomposition of X that corresponds to the
covariates of interest. May be removed in the future.}

\item{degrees_freedom}{A positive numeric. The degrees of freedom
of the t-likelihood if using it.}

\item{gls}{A logical. Should we use generalized least squares
(\code{TRUE}) or ordinary least squares (\code{FALSE}) for
estimating the confounders? The OLS version is equivalent to
using RUV4 to estimate the confounders.}

\item{likelihood}{Either \code{"normal"} or \code{"t"}. If
\code{likelihood = "t"}, then the user may provide the degrees
of freedom via \code{degrees_freedom}.}
}
\description{
RUV4's second step.
}
\author{
David Gerard
}

