% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash_wrap.R
\name{dnormalmix}
\alias{dnormalmix}
\title{Density of a mixture of normals}
\usage{
dnormalmix(x, mixdense)
}
\arguments{
\item{x}{The quantile.}

\item{mixdense}{An object of class \code{\link[ashr]{normalmix}}.}
}
\description{
Density of a mixture of normals
}
\author{
David Gerard
}

