% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash.R
\name{dt_wrap}
\alias{dt_wrap}
\title{Wrapper for dt with a non-zero mean and non-1 scale parameter.}
\usage{
dt_wrap(x, df, mean = 0, sd = 1, log = FALSE)
}
\arguments{
\item{x}{A numeric. Where to evaluate the density function.}

\item{df}{A positive numeric. The degrees of freedom.}

\item{mean}{A numeric. The mean of the t. Defaults to 0.}

\item{sd}{A positive numeric. The standard deviation of the
t. Defaults to 1.}

\item{log}{A logical. Should we return the log-density
(\code{TRUE}) or the density (\code{FALSE})?}
}
\description{
Wrapper for dt with a non-zero mean and non-1 scale parameter.
}
