% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_miss.R
\name{em_miss_fix_fast}
\alias{em_miss_fix_fast}
\title{Faster version of \code{\link{em_miss_fix}}.}
\usage{
em_miss_fix_fast(alpha_sigma, Y21, Y31, Y32, k)
}
\arguments{
\item{alpha_sigma}{A vector. The first k * p elements are the
vectorization of alpha. The last elements are sig_diag.}

\item{Y21}{The top left matrix.}

\item{Y31}{The bottom left matrix.}

\item{Y32}{The bottom right matrix.}

\item{k}{The rank of the mean.}
}
\description{
Faster version of \code{\link{em_miss_fix}}.
}
\author{
David Gerard
}
