% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_miss.R
\name{em_miss_obj}
\alias{em_miss_obj}
\title{The objective function for em with a missing block.}
\usage{
em_miss_obj(alpha_sigma, S, S2, ncovs, ncontrols, n, p, k)
}
\arguments{
\item{alpha_sigma}{A vector. The first k * p elements are the
vectorization of alpha. The last elements are sig_diag.}

\item{S}{A matrix. This is the sample covariance matrix of the
samples that don't have a missing block.}

\item{S2}{A matrix. This is the sample covariance matrix of the
samples that do have a missing block.}

\item{ncovs}{The number of covariates.}

\item{ncontrols}{The number of controls}

\item{n}{The sample size}

\item{p}{The number of genes.}

\item{k}{The rank of the mean.}
}
\description{
The objective function for em with a missing block.
}
\author{
David Gerard
}
