% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_miss.R
\name{em_miss_obj_fast}
\alias{em_miss_obj_fast}
\title{A faster version of em_miss_obj.}
\usage{
em_miss_obj_fast(alpha_sigma, Y21, Y31, Y32, k)
}
\arguments{
\item{alpha_sigma}{A vector. The first k * p elements are the
vectorization of alpha. The last elements are sig_diag.}

\item{Y21}{The top left matrix.}

\item{Y31}{The bottom left matrix.}

\item{Y32}{The bottom right matrix.}

\item{k}{The rank of the mean.}
}
\description{
A faster version of em_miss_obj.
}
\author{
David Gerard
}
\seealso{
\code{\link{em_miss_fix_fast}} for a fixed point iteration
    that will maximize this function. \code{\link{em_miss_obj}} for
    the slower version of this function.
}

