% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv4.R
\name{fa_ml}
\alias{fa_ml}
\title{Wrapper for cate's fa.em function.}
\usage{
fa_ml(Y, r)
}
\arguments{
\item{Y}{A matrix of numerics. The data.}

\item{r}{the rank.}
}
\description{
Wrapper for cate's fa.em function.
}
\references{
Jingshu Wang and Qingyuan Zhao (2015). cate: High
    Dimensional Factor Analysis and Confounder Adjusted Testing and
    Estimation. R package version
    1.0.4. https://CRAN.R-project.org/package=cate
}
\author{
David Gerard
}
