% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_tester.R
\name{fa_out_test}
\alias{fa_out_test}
\title{Test output form fa_func.}
\usage{
fa_out_test(fa_out, fa_args)
}
\arguments{
\item{fa_out}{Output from fa_func.}

\item{fa_args}{Input to fa_func.}
}
\description{
Test output form fa_func.
}
\author{
David Gerard
}
