% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_tester.R
\name{fa_tester}
\alias{fa_tester}
\title{Tests whether a user-specified factor analysis function
is appropriate for use in \code{\link{vruv4}}, \code{\link{ruv3}},
\code{\link{mouthwash}}, or \code{\link{backwash}}.}
\usage{
fa_tester(fa_func, fa_args = list(), test_ortho = TRUE)
}
\arguments{
\item{fa_func}{The factor analysis function to be tested.}

\item{fa_args}{A list of additional arguments to pass to \code{fa_func}.}

\item{test_ortho}{A logical. Should we test for left orthogonal equivariance (\code{TRUE}),
or not (\code{FALSE})?}
}
\value{
A list with some or all of the following elements:

    \code{ok} A logical. A value of \code{TRUE} means the function is OK. A value of
    \code{FALSE} means the function is not OK.

    \code{why} A string vector giving the reason(s) why the function is not OK if \code{ok = FALSE}.

    \code{input} A list of input values to help you debug.
}
\description{
This function tests that the output of \code{fa_func} is compatible with the
confounder adjustment methods in this package. In general, \code{fa_func} needs to take as input
at least a matrix \code{Y}, which is dimensions n by p, and the rank of the factor analysis \code{r}.
\code{fa_func} needs to return three elements, a matrix \code{alpha} that is p by r, a matrix \code{Z} that
is n by r, and a numeric vector \code{sig_diag} that is of length \code{p}.
}
\details{
This function also tests for orthogonal equivariance. See Gerard and Stephens (2016) for a definition
of left orthogonal equivariance.
You can turn off this test with \code{test_ortho = FALSE}, but any factor analysis that is not
left orthogonally equivariant would be dubious as the rotation onto the null space is completely
arbitrary.
}
\author{
David Gerard
}
