% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_tester.R
\name{fa_tester_ruvb}
\alias{fa_tester_ruvb}
\title{Test whether a user-specified function is compatible with \code{\link{ruvb}}.}
\usage{
fa_tester_ruvb(fa_func, fa_args = list())
}
\arguments{
\item{fa_func}{The factor analysis function.}

\item{fa_args}{Additional arguments to pass to fa_func.}
}
\value{
A list with some or all of the following elements:

    \code{ok} A logical. A value of \code{TRUE} means the function is OK. A value of
    \code{FALSE} means the function is not OK.

    \code{why} A string vector giving the reason(s) why the function is not OK if \code{ok = FALSE}.

    \code{input} A list of input values to help you debug.
}
\description{
This function tests whether a user-defined function \code{fa_func} is compatible
for use in \code{\link{ruvb}}. Let n be the number of samples, p be the number of genes,
m be the number of control genes, and q be the number of covariates.
In general, \code{fa_func} needs to take as input
\code{Y21} a q by m matrix of numerics, \code{Y31} a (n - q) by m matrix of numerics,
\code{Y32} a (n - q) by (p - m) matrix of numerics, and \code{k} a positive integer
that is the rank of the factor model if assuming a factor model.
\code{fa_func} also needs to return \code{Y22_array} a three-way array whose first two dimensions are
k and (p - m).
}
\seealso{
\code{\link{ruvb}}
}
\author{
David Gerard
}
