% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caw.R
\name{fix_caw}
\alias{fix_caw}
\title{A fixed point iteration for CAW.}
\usage{
fix_caw(
  w_current,
  Z2hat,
  scale_val,
  rotate_out,
  weight_func = ash_wrap,
  weight_args = list(),
  min_scale = 0.8,
  return_val = c("params", "full")
)
}
\arguments{
\item{w_current}{A vector of weights. The probability given Z2hat
that each coefficient is null. These should all be between 0
and 1.}

\item{Z2hat}{A matrix. The current estimates of the unobserved
confounders.}

\item{scale_val}{A positive numeric. The variance inflation term.}

\item{rotate_out}{The output from \code{\link{rotate_model}} with
\code{degrees_freedom} added to the list.}

\item{weight_func}{The function that calculates the weights given
\code{w_current}. So far, this only supports functions that
take as input three quantities: \code{betahat},
\code{sebetahat}, and \code{degrees_freedom}}

\item{weight_args}{Additional arguments to pass to
\code{weight_func}.}

\item{min_scale}{A scalar. The minimum value that \code{scale_val}
is allowed to take. In practice, you wouldn't think that
shrinking the variance is a good idea. But a value of 1 seems
to stringent. So I set this to 0.8 to give you some wiggle
room. Most of the time \code{scale_val} is usually estimated
greater than 1.}

\item{return_val}{Should we return just the parameters of interest
(\code{"params"}), or the full shebang (\code{"full"})? Setting
\code{return_val = "params"} is used mostly for the
optimization and returns only \code{w_current}, \code{Z2hat},
and \code{scale_val}.}
}
\description{
A fixed point iteration for CAW.
}
\author{
David Gerard
}
