% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caw.R
\name{fix_caw_wrapper}
\alias{fix_caw_wrapper}
\title{A wrapper for \code{\link{fix_caw}} so that I can use SQUAREM.}
\usage{
fix_caw_wrapper(
  wzs_vec,
  rotate_out,
  weight_func = ash_wrap,
  weight_args = list(),
  min_scale = 0.8
)
}
\arguments{
\item{wzs_vec}{A vector of numerics. The first p of which are the
\code{w_current}, the next \code{length(cov_of_interest) * k}
of which are \code{Z2hat}, the last of which is
\code{scale_val}.}

\item{rotate_out}{The output from \code{\link{rotate_model}} with
\code{degrees_freedom} added to the list.}

\item{weight_func}{The function that calculates the weights given
\code{w_current}. So far, this only supports functions that
take as input three quantities: \code{betahat},
\code{sebetahat}, and \code{degrees_freedom}}

\item{weight_args}{Additional arguments to pass to
\code{weight_func}.}

\item{min_scale}{A scalar. The minimum value that \code{scale_val}
is allowed to take. In practice, you wouldn't think that
shrinking the variance is a good idea. But a value of 1 seems
to stringent. So I set this to 0.8 to give you some wiggle
room. Most of the time \code{scale_val} is usually estimated
greater than 1.}
}
\description{
A wrapper for \code{\link{fix_caw}} so that I can use SQUAREM.
}
\author{
David Gerard
}
