% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{gdfa}
\alias{gdfa}
\title{Old version of Bayesian factor analysis.}
\usage{
gdfa(Y21, Y31, Y32, k, nsamp = 10000, burnin = round(nsamp/4),
  keep = 20, print_update = TRUE, plot_update = FALSE,
  hetero_factors = FALSE, rho_0 = 0.1, alpha_0 = 0.1,
  delta_0 = 0.1, lambda_0 = 0.1, nu_0 = 1, beta_0 = 1, eta_0 = 1,
  tau_0 = 1)
}
\arguments{
\item{Y21}{Top left of matrix.}

\item{Y31}{Bottom left of matrix.}

\item{Y32}{Top right of matrix.}

\item{k}{The rank of the mean matrix.}

\item{nsamp}{A positive integer. The number of samples to draw.}

\item{burnin}{A positive integer. The number of early samples to
discard.}

\item{keep}{A positive integer. We will same the updates of
\code{Y22} every \code{keep} iteration of the Gibbs sampler.}

\item{print_update}{A logical. Should we print a text progress bar
to keep track of the Gibbs sampler (\code{TRUE}) or not
(\code{FALSE})?}

\item{plot_update}{A logical. Should we make some plots to keep
track of the Gibbs sampler (\code{TRUE}) or not (\code{FALSE})?}

\item{hetero_factors}{A logical. Should we assign colum-specific
variances for the factors (\code{TRUE}) or not (\code{FALSE})?}

\item{rho_0}{The prior sample size for column-specific the
precisions.}

\item{alpha_0}{The prior sample size for the mean of the
column-specific precisions.}

\item{delta_0}{The prior sample size of the column-specific
precisions of the factors.}

\item{lambda_0}{The prior sample size of the mean of the
column-specific precisions of the factors.}

\item{nu_0}{The prior mean of the mean of the column-specific
precisions of the factors.}

\item{beta_0}{The prior mean of the mean of the column-specific
precisions.}

\item{eta_0}{The prior sample size of the expanded parameters.}

\item{tau_0}{The prior mean of of the expanded parameters.}
}
\description{
See \code{\link{bfa_gs}} for a description. This is the same thing
but a slower R implementation.
}
\references{
\itemize{
\item{Ghosh, Joyee, and David
    B. Dunson. "Default prior distributions and efficient posterior
    computation in Bayesian factor analysis."
    Journal of Computational and Graphical Statistics 18.2 (2009):
    306-320.}
}
}
\seealso{
\code{\link{bfa_gs}}
}
\author{
David Gerard
}
