% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash.R
\name{get_grid_var}
\alias{get_grid_var}
\title{Default way to set grid of variances.}
\usage{
get_grid_var(betahat_ols, S_diag)
}
\arguments{
\item{betahat_ols}{The OLS estimates of the coefficients of interest.}

\item{S_diag}{The standard error estimates.}
}
\description{
Default way to set grid of variances.
}
\author{
David Gerard
}
