% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv3.R
\name{hard_impute}
\alias{hard_impute}
\title{My version of hard imputation that begins at the ruv estimates}
\usage{
hard_impute(Y21, Y31, Y32, k, tol = 10^-5, maxit = 1000,
  init_type = c("ruv4", "ruv3"))
}
\arguments{
\item{Y21}{Top left of matrix.}

\item{Y31}{Bottom left of matrix.}

\item{Y32}{Top right of matrix.}

\item{k}{The rank of the mean matrix.}

\item{tol}{The tolerance for stopping.}

\item{maxit}{The maximum number of iterations to run.}

\item{init_type}{Which version of RUV should we start?
\code{"ruv3"} or \code{"ruv4"}?}
}
\value{
The top right of the matrix.
}
\description{
My version of hard imputation that begins at the ruv estimates
}
\author{
David Gerard
}

