% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvimpute.R
\name{impute_block}
\alias{impute_block}
\title{Constructs an overall matrix, then applies a given imputation function.}
\usage{
impute_block(Y21, Y31, Y32, impute_func, impute_args = list(),
  do_variance = FALSE)
}
\arguments{
\item{Y21}{A matrix of size k by m}

\item{Y31}{A matrix of size (n - k) by m}

\item{Y32}{A matrix of size (n - k) by (p - m)}

\item{impute_func}{A function that takes as input a matrix names
\code{Y} that has missing values and returns a matrix called
\code{Yhat} of the same dimension of \code{Y} with the missing
values filled in. If \code{do_variance = TRUE}, then
\code{impute_func} should also return \code{sig_diag} --- a
vector of column-specific variance estimates. The default is a
wrapper for \code{\link[softImpute]{softImpute}}.  I provide a
few functions in this package. \code{\link{knn_wrapper}}
performs k-nearest-neighbors
imputation. \code{\link{missforest_wrapper}} performs random
forest imputation. \code{\link{softimpute_wrapper}} performs
nuclear norm minimization imputation.}

\item{impute_args}{A list of additional parameters to pass to
\code{impute_func}.}

\item{do_variance}{A logical. Does \code{impute_func} also return
estimates of the column-specific variances?}
}
\value{
\code{Y22} A matrix of size k by (p - m). This is the
    imputed submatrix.

    \code{sig_diag} If \code{do_variance = FALSE}, then this is
    \code{NULL}. Else, it is a vector of column-specific variance
    estimates.
}
\description{
Constructs an overall matrix, then applies a given imputation function.
}
\author{
David Gerard
}

