% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvimpute.R
\name{impute_ruv_reproduce}
\alias{impute_ruv_reproduce}
\title{Reproduce RUV2, RUV3, and RUV4 with RUVimpute.}
\usage{
impute_ruv_reproduce(Y21, Y31, Y32, k, impute_type = c("ruv2", "ruv3",
  "ruv4"))
}
\arguments{
\item{Y21}{Top left of matrix.}

\item{Y31}{Bottom left of matrix.}

\item{Y32}{Top right of matrix.}

\item{k}{The rank of the mean matrix.}

\item{impute_type}{Which version of RUV should we reproduce?
\code{"ruv2"}, \code{"ruv3"}, or \code{"ruv4"}?}
}
\description{
This is a proof of concept function to inegrate
\code{\link{ruvimpute}} with \code{\link[ruv]{RUV2}},
\code{\link{ruv3}}, \code{\link[ruv]{RUV4}}.
}
\author{
David Gerard
}
