% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash.R
\name{initialize_mixing_prop}
\alias{initialize_mixing_prop}
\title{Function for initializing mixing proportions.}
\usage{
initialize_mixing_prop(pi_init_type = c("zero_conc", "uniform", "random"),
  zero_spot, M)
}
\arguments{
\item{pi_init_type}{How should we initialize the mixing
proportions? By concentrating on zero (\code{"zero_conc"}), by
equal weights on all mixing distributions (\code{"uniform"}),
or by sampling uniformly on the simplex (\code{"random"})?}

\item{zero_spot}{The location of the zero.}

\item{M}{The number of mixing distributions.}
}
\description{
Function for initializing mixing proportions.
}
\author{
David Gerard
}
