% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapp.R
\name{leapp_cross}
\alias{leapp_cross}
\title{10-fold cross-validation-ish method.}
\usage{
leapp_cross(X, Y, sigma, betaInit, method = "hard", TOL = 1e-04,
  num_sets = 10)
}
\description{
10-fold cross-validation-ish method.
}

