% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv3.R
\name{missforest_wrapper}
\alias{missforest_wrapper}
\title{Wrapper for missForest package.}
\usage{
missforest_wrapper(Y)
}
\arguments{
\item{Y}{A matirx with missing values.}
}
\value{
A matrix with the missing values imputed.
}
\description{
Wrapper for missForest package.
}
\author{
David Gerard
}
\seealso{
\code{\link[missForest]{missForest}}
}

