% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash_uniform_coordinate.R
\name{mouthwash_coordinate}
\alias{mouthwash_coordinate}
\title{Coordinate ascent for optimizing t likelihood with uniform mixtures.}
\usage{
mouthwash_coordinate(pi_init, z_init, xi_init, betahat_ols, S_diag,
  alpha_tilde, a_seq, b_seq, lambda_seq, degrees_freedom,
  scale_var = TRUE, tol = 10^-6, maxit = 100, plot_update = FALSE,
  var_inflate_pen = 0)
}
\arguments{
\item{pi_init}{The initial values of the mixing proportions. A
numeric vector of positive values that sum to one.}

\item{z_init}{The initial values of z2. A numeric vector.}

\item{xi_init}{The initial (or known) value of the variance
inflation term.}

\item{betahat_ols}{A vector of numerics. The OLS estimates of the
coefficients of interest.}

\item{S_diag}{A vector of positive numerics. The estimated standard
errors.}

\item{alpha_tilde}{A matrix. The number of rows should be equal the
length of betahat_ols. The number of columns should equal the
number of hidden confounders.}

\item{a_seq}{The grid of lower bounds for the uniforms if
\code{mixing_dist} is one of the uniforms.}

\item{b_seq}{The grid of upper bounds for the uniforms if
\code{mixing_dist} is one of the uniforms.}

\item{lambda_seq}{A numeric vector with elements all greater than
or equal to 1. These are the tuning parameters for the mixing
proportions.}

\item{degrees_freedom}{The degrees of freedom of the t-distribution
if \code{likelihood = "t"}.}

\item{scale_var}{A logical. Should we estimate a variance inflation
parameter (\code{TRUE}) or not (\code{FALSE})?}

\item{tol}{The tolerance for the stopping condition.}

\item{maxit}{The maximum number of iterations to run in the
optimization scheme.}

\item{plot_update}{A logical. Should I plot the the path of the
log-likelihood (\code{TRUE}) or not (\code{FALSE})?}

\item{var_inflate_pen}{The penalty to apply on the variance inflation parameter.
Defaults to 0, but should be something non-zero when \code{alpha = 1}
and \code{scale_var = TRUE}.}
}
\description{
Set \code{degrees_freedom = Inf} if you want to use a normal likelihood.
}
\author{
David Gerard
}
