% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash_uniform_coordinate.R
\name{mouthwash_z_grad}
\alias{mouthwash_z_grad}
\title{Gradient wrt z of \code{\link{uniform_mix_llike}}.}
\usage{
mouthwash_z_grad(pi_vals, z2, xi, betahat_ols, S_diag, alpha_tilde, a_seq,
  b_seq, lambda_seq, degrees_freedom)
}
\description{
Gradient wrt z of \code{\link{uniform_mix_llike}}.
}
\author{
David Gerard
}
\seealso{
\code{\link{uniform_mix_llike}} for the objective function
    and \code{\link{mouthwash_coordinate}} for where this function
    is used.
}

