% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash_normal_em.R
\name{normal_mix_fix}
\alias{normal_mix_fix}
\title{A fixed point iteration for updating the mixing proportions and the
confounders associated with the covariates of interest when using a
mixture of normals prior.}
\usage{
normal_mix_fix(pi_vals, z2, xi, betahat_ols, S_diag, alpha_tilde, tau2_seq,
  lambda_seq, scale_var = TRUE, var_inflate_pen = 0)
}
\arguments{
\item{pi_vals}{The current values of the mixing proportions.}

\item{z2}{The current value of the unobserved confounders
corresponding to the covariates of interest.}

\item{xi}{The current value of the variance inflation parameter.}

\item{betahat_ols}{A vector of numerics. The OLS estimates of the
coefficients of interest.}

\item{S_diag}{A vector of positive numerics. The estimated standard
errors.}

\item{alpha_tilde}{A matrix. The number of rows should be equal the
length of betahat_ols. The number of columns should equal the
number of hidden confounders.}

\item{tau2_seq}{The grid of variances. This is the same thing as
grid_seq in \code{\link{mouthwash_second_step}} if
\code{likelihood = "normal"}.}

\item{lambda_seq}{A numeric vector with elements all greater than
or equal to 1. These are the tuning parameters for the mixing
proportions.}

\item{scale_var}{Should we optimize over a variance inflation
parameter (\code{TRUE}) or not (\code{FALSE})?}

\item{var_inflate_pen}{The penalty to apply on the variance inflation parameter.
Defaults to 0, but should be something non-zero when \code{alpha = 1}
and \code{scale_var = TRUE}.}
}
\value{
A list with the following elements.

    \code{pi_vals}: The update for \code{pi_vals}.

    \code{z2}: The update for \code{z2}.

    \code{xi}: The update for \code{xi}.
}
\description{
A fixed point iteration for updating the mixing proportions and the
confounders associated with the covariates of interest when using a
mixture of normals prior.
}
\author{
David Gerard
}
