% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash.R
\name{normal_mix_fix_wrapper}
\alias{normal_mix_fix_wrapper}
\title{Wrapper for \code{\link{normal_mix_fix}} so that I can use SQUAREM.}
\usage{
normal_mix_fix_wrapper(pizxi_vec, betahat_ols, S_diag, alpha_tilde, tau2_seq,
  lambda_seq, scale_var = TRUE)
}
\arguments{
\item{pizxi_vec}{A vector whose first M elements are pi_vals, next
k elements are z2, and last element is xi.}

\item{betahat_ols}{A vector of numerics. The OLS estimates of the
coefficients of interest.}

\item{S_diag}{A vector of positive numerics. The estimated standard
errors.}

\item{alpha_tilde}{A matrix. The number of rows should be equal the
length of betahat_ols. The number of columns should equal the
number of hidden confounders.}

\item{tau2_seq}{The grid of variances. This is the same thing as
grid_seq in \code{\link{mouthwash_second_step}} if
\code{likelihood = "normal"}.}

\item{lambda_seq}{A numeric vector with elements all greater than
or equal to 1. These are the tuning parameters for the mixing
proportions. This can only be specified if \code{grid_seq} is
also specified.}

\item{scale_var}{Should we optimize over a variance inflation
parameter (\code{TRUE}) or not (\code{FALSE})?}
}
\description{
Wrapper for \code{\link{normal_mix_fix}} so that I can use SQUAREM.
}
\author{
David Gerard
}
\seealso{
\code{\link{normal_mix_fix}}
}

