% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash.R
\name{normal_mix_llike}
\alias{normal_mix_llike}
\title{Penalized MOUTHWASH likelihood when using a mixture of normals.}
\usage{
normal_mix_llike(pi_vals, z2, xi, betahat_ols, S_diag, alpha_tilde, tau2_seq,
  lambda_seq)
}
\arguments{
\item{pi_vals}{The current values of the mixing proportions.}

\item{z2}{The current value of the unobserved confounders
corresponding to the covariates of interest.}

\item{xi}{The current value of the variance inflation parameter.}

\item{betahat_ols}{A vector of numerics. The OLS estimates of the
coefficients of interest.}

\item{S_diag}{A vector of positive numerics. The estimated standard
errors.}

\item{alpha_tilde}{A matrix. The number of rows should be equal the
length of betahat_ols. The number of columns should equal the
number of hidden confounders.}

\item{tau2_seq}{The grid of variances. This is the same thing as
grid_seq in \code{\link{mouthwash_second_step}} if
\code{likelihood = "normal"}.}

\item{lambda_seq}{A numeric vector with elements all greater than
or equal to 1. These are the tuning parameters for the mixing
proportions. This can only be specified if \code{grid_seq} is
also specified.}
}
\description{
Penalized MOUTHWASH likelihood when using a mixture of normals.
}
\author{
David Gerard
}

