% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv2_fa.R
\name{pca_2step}
\alias{pca_2step}
\title{PCA when first \code{vr} rows have a variance multiplicatively
different from the rest of the rows.}
\usage{
pca_2step(Y, r, vr, limmashrink = TRUE, likelihood = c("t", "normal"))
}
\arguments{
\item{Y}{A matrix of numerics. The data.}

\item{r}{the rank.}

\item{vr}{The number of the first few rows whose variances differ
by a multiplicative factor.}

\item{limmashrink}{A logical. Should we shrink the variance
estimates prior to performing gls to get Z1 (\code{TRUE}) or
not (\code{FALSE})?}

\item{likelihood}{What should the likelihood be when we estimate
Z1? Options are \code{"t"} and \code{"normal"}.}
}
\description{
Modified truncated SVD. The variance estimates are just the
column-wise mean-squares of the last \eqn{n - vr} rows. This form
of factor analysis is mostly for variance inflation with RUV2.
}
\details{
This doesn't work too well. I think the Z's and sigmas need to be
estimated jointly.
}
\author{
David Gerard
}
