% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv_ash.R
\name{pca_naive}
\alias{pca_naive}
\title{Basic PCA.}
\usage{
pca_naive(Y, r)
}
\arguments{
\item{Y}{A matrix of numerics. The data.}

\item{r}{the rank.}
}
\description{
Glorified truncated SVD. The variance estimates are just the
column-wise mean-squares, except I divide by \code{nrow(Y) - r}
rather than by \code{nrow(Y)}. I allow \code{r = 0}.
}
\author{
David Gerard
}

