% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv2_fa.R
\name{pca_ruv2}
\alias{pca_ruv2}
\title{PCA when first \code{vr} rows have a variance multiplicatively
different from the rest of the rows.}
\usage{
pca_ruv2(Y, r, vr, mle = FALSE)
}
\arguments{
\item{Y}{A matrix of numerics. The data.}

\item{r}{the rank.}

\item{vr}{The number of the first few rows whose variances differ
by a multiplicative factor.}

\item{mle}{A logical. Should we run an MLE on the residuals of PCA
(\code{TRUE}) or just use a two-step estimator (\code{FALSE}).}
}
\description{
Modified truncated SVD. The variance estimates are just the
column-wise mean-squares of the last \eqn{n - vr} rows. This form
of factor analysis is mostly for variance inflation with RUV2.
}
\details{
This doesn't work too well. I think the Z's and sigmas need to be
estimated jointly.
}
\author{
David Gerard
}
