% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv2.R
\name{pcaruv2_fix}
\alias{pcaruv2_fix}
\title{Fix point for mle in pca_ruv2.}
\usage{
pcaruv2_fix(sig_lambda, r1, r2, n, vr)
}
\arguments{
\item{sig_lambda}{A vector. All but the first elements are the
current values of Sigma. The last element is the current valur
of lambda.}

\item{r1}{The sum of squares of the first vr columns of the
residuals.}

\item{r2}{The sum of squares of the last n - vr columns of the
residuals.}

\item{n}{The number of samples.}

\item{vr}{The number of columns where variance inflation is in
effect.}
}
\description{
This is mostly for use in SQUAREM.
}
\author{
David Gerard
}
\seealso{
\code{\link{pca_ruv2}}, \code{\link{pcaruv2_obj}} for the
    objective function that this maximizes.
}

