% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.backwash}
\alias{plot.backwash}
\title{Plotting method for \code{\link{backwash}}.}
\usage{
\method{plot}{backwash}(x, ...)
}
\arguments{
\item{x}{An object of class \code{backwash}, returned from \code{\link{backwash}}.}

\item{...}{Not used.}
}
\description{
Plotting method for \code{\link{backwash}}.
}
\author{
David Gerard
}
