% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.mouthwash}
\alias{plot.mouthwash}
\title{Plotting method for \code{moutwash}.}
\usage{
\method{plot}{mouthwash}(x, ...)
}
\arguments{
\item{x}{The output of \code{\link{mouthwash}}, of class \code{mouthwash}.}

\item{...}{Not used.}
}
\description{
This will produce
}
\seealso{
\code{\link{mouthwash}}.
}
\author{
David Gerard
}
