% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.ruvb}
\alias{plot.ruvb}
\title{Plotting method for \code{ruvb}.}
\usage{
plot.ruvb(x, ...)
}
\arguments{
\item{x}{The output of \code{\link{ruvb}}, of class \code{ruvb}.}

\item{...}{Not used.}
}
\description{
For the traceplots, the blue dashed line is a non-parameteric smoother.
This should let you see if there are any major trends in the traceplot.
}
\seealso{
\code{\link{ruvb}}
}
\author{
David Gerard
}
