% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{prior_fun_wrapper}
\alias{prior_fun_wrapper}
\title{Wrapper for \code{prior_fun} so that can be called in apply.}
\usage{
prior_fun_wrapper(beta_mat, prior_fun, prior_args = list())
}
\arguments{
\item{beta_mat}{A matrix}

\item{prior_fun}{A function.}

\item{prior_args}{A list of arguments}
}
\description{
Wrapper for \code{prior_fun} so that can be called in apply.
}
\author{
David Gerard
}

