% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash.R
\name{pt_wrap}
\alias{pt_wrap}
\title{Wrapper for pt with a non-zero mena and non-1 scale parameter.}
\usage{
pt_wrap(x, df, mean = 0, sd = 1)
}
\arguments{
\item{x}{A numeric. Where to evaluate the cdf.}

\item{df}{A positive numeric. The degrees of freedom.}

\item{mean}{A numeric. The mean of the t. Defaults to 0.}

\item{sd}{A positive numeric. The standard deviation of the
t. Defaults to 1.}
}
\description{
Wrapper for pt with a non-zero mena and non-1 scale parameter.
}

