% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash_uniform_em.R
\name{ptdiff}
\alias{ptdiff}
\title{More stable way to calculate differences in T cdf's.}
\usage{
ptdiff(x, y, degrees_freedom)
}
\arguments{
\item{x}{A quantile of the t.}

\item{y}{The other quantile of the t.}

\item{degrees_freedom}{The degrees of freedom of the t.}
}
\description{
If \eqn{T()} is the cdf of a t-distribution if
\code{degrees_freedom} degrees of freedom, then this function will
calculate \eqn{T(x) - T(y)} or \eqn{T(-y) - T(-x)} (which are are
equivalent), depending on which one is more numerically stable.
}
\author{
David Gerard
}

