% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash_uniform_em.R
\name{ptdiff_mat_log}
\alias{ptdiff_mat_log}
\title{Log version of \code{\link{ptdiff_mat}}.}
\usage{
ptdiff_mat_log(X, Y, degrees_freedom)
}
\arguments{
\item{X}{A matirx of quantiles of the t.}

\item{Y}{Another matirx of quantiles of the t.}

\item{degrees_freedom}{The degrees of freedom of the t.}
}
\description{
Log version of \code{\link{ptdiff_mat}}.
}
\seealso{
\code{\link{ptdiff_mat}}.
}
\author{
David Gerard
}
