% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv2_fa.R
\name{qmle_obj}
\alias{qmle_obj}
\title{Objective function for quasi-mle approach.}
\usage{
qmle_obj(alpha, sig_diag, lambda, Y1, Y2)
}
\arguments{
\item{alpha}{The current factor loadings}

\item{sig_diag}{A vector of the current variances}

\item{lambda}{The current inflation parameter for Y1}

\item{Y1}{The first part of the observations.}

\item{Y2}{The second part of the observations.}
}
\description{
This isn't really the log-likelihood, but the log-likelihood plus
some constant.
}
\details{
For separate parts of hte likelihood, this code borrows heavily
from the function \code{\link[cate]{fa.em}}.
}
