% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv2.R
\name{qmle_obj_basic}
\alias{qmle_obj_basic}
\title{Basic no optimized code objective function.}
\usage{
qmle_obj_basic(alpha, sig_diag, lambda, Y1, Y2)
}
\arguments{
\item{alpha}{The current factor loadings}

\item{sig_diag}{A vector of the current variances}

\item{lambda}{The current inflation parameter for Y1}

\item{Y1}{The first part of the observations.}

\item{Y2}{The second part of the observations.}
}
\description{
Basic no optimized code objective function.
}
\author{
David Gerard
}

