% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv2_fa.R
\name{qmle_ruv2}
\alias{qmle_ruv2}
\title{Quasi-mle when first \code{vr} rows of \code{Y} have variances
which differ by a scale factor from the rest of the rows.}
\usage{
qmle_ruv2(Y, r, vr, tol = 10^-3, maxit = 1000)
}
\arguments{
\item{Y}{A matrix of numerics. The data.}

\item{r}{the rank.}

\item{vr}{The number of the first few rows whose variances differ
by a multiplicative factor.}

\item{tol}{The function tolerance for stopping. A positive numeric.}

\item{maxit}{The maximum number of iterations to run.}
}
\value{
\code{alpha} A matrix of the estimated factor loadings.

        \code{Z} A matrix of the estimated factors.

        \code{sig_diag} A vector of the estimated column-wise
        variances.

        \code{lambda} A numeric scalar of the estimated variance
        inflation parameter.
}
\description{
Quasi-mle when first \code{vr} rows of \code{Y} have variances
which differ by a scale factor from the rest of the rows.
}
\author{
David Gerard
}
