% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv2_fa.R
\name{qmle_ruv2_lambda_grid}
\alias{qmle_ruv2_lambda_grid}
\title{Quasi-mle when first \code{vr} rows of \code{Y} have variances
which differ by a scale factor from the rest of the rows.}
\usage{
qmle_ruv2_lambda_grid(Y, r, vr, gridsize = 20, gridlow = 1, gridhigh = 10,
  tol = 10^-3, maxit = 1000, plot_llike = FALSE)
}
\arguments{
\item{Y}{A matrix of numerics. The data.}

\item{r}{the rank.}

\item{vr}{The number of the first few rows whose variances differ
by a multiplicative factor.}

\item{gridsize}{A positive integer. The size of the grid for the
variance inflation parameter.}

\item{gridlow}{A positive numeric. The minimum variance inflation
parameter to try out.}

\item{gridhigh}{A positive numeric. The maximum variance inflation
parameter to try out.}

\item{tol}{The function tolerance for stopping. A positive numeric.}

\item{maxit}{The maximum number of iterations to run.}

\item{plot_llike}{A logical. Should I plot the profiled
log-likelihood?}
}
\value{
\code{alpha} A matrix of the estimated factor loadings.

        \code{Z} A matrix of the estimated factors.

        \code{sig_diag} A vector of the estimated column-wise
        variances.

        \code{lambda} A numeric scalar of the estimated variance
        inflation parameter.
}
\description{
This one will iterate through a grid of lambdas and choose the one
with the highest likelihood.
}
\author{
David Gerard
}
