% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv4.R
\name{qr_ident}
\alias{qr_ident}
\title{An identified QR decomposition.}
\usage{
qr_ident(X)
}
\arguments{
\item{X}{A matrix.}
}
\description{
This just re-jiggers the output from \code{\link[base]{qr}} so that
the upper triangular matrix has positive diagonal elements. This
one will only work if the column dimension is less than or equal to
the row dimension.
}

