% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash_wrap.R
\name{rmixnorm}
\alias{rmixnorm}
\title{Draw from a mixture of normals.}
\usage{
rmixnorm(n, pi_vals, sd_seq, mean_seq = rep(0, length(pi_vals)))
}
\arguments{
\item{n}{An integer. The number of samples to draw.}

\item{pi_vals}{A vector length \code{M}. The mixing proportions.}

\item{sd_seq}{A vector of length \code{M}. The mixing standard
deviations.}

\item{mean_seq}{A vector of length \code{M}. The mixing means.}
}
\value{
A vector of length \code{p} that are drawn from the mean
    zero mixture of normals.
}
\description{
Draw from a mean zero mixture of normals given mixing proportions,
mixing standard deviations, and mixing means.
}
\details{
Given mixing proportions \code{pi_vals}, and mixing standard
deviations (not variances) \code{sd_seq}, this function will draw
\code{p} samples from the mean zero mixture of normals.
}
