% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash_wrap.R
\name{rnormalmix}
\alias{rnormalmix}
\title{Random draw from a mixture of normals.}
\usage{
rnormalmix(n, mixdense)
}
\arguments{
\item{n}{the number of samples to draw.}

\item{mixdense}{An object of class \code{\link[ashr]{normalmix}}.}
}
\description{
Random draw from a mixture of normals.
}
\author{
David Gerard
}
