% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_miss.R
\name{rubin_copy}
\alias{rubin_copy}
\title{Very inneficient copy of rubin and thayer iteration mostly meant for debugging.}
\usage{
rubin_copy(alpha_sigma, S, n, p, k)
}
\arguments{
\item{alpha_sigma}{A vector. The first k * p elements are the
vectorization of alpha. The last elements are sig_diag.}

\item{S}{A matrix. This is the sample covariance matrix of the
samples that don't have a missing block.}

\item{n}{The sample size}

\item{p}{The number of genes.}

\item{k}{The rank of the mean.}
}
\description{
Very inneficient copy of rubin and thayer iteration mostly meant for debugging.
}
\author{
David Gerard
}
