% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv3.R
\name{ruv3}
\alias{ruv3}
\title{Removing Unwanted Variation 3.}
\usage{
ruv3(Y, X, ctl, k = NULL, cov_of_interest = ncol(X),
  include_intercept = TRUE, limmashrink = TRUE, gls = TRUE,
  fa_func = pca_naive, fa_args = list())
}
\arguments{
\item{Y}{A matrix of numerics. These are the response variables
where each column has its own variance. In a gene expression
study, the rows are the individuals and the columns are the
genes.}

\item{X}{A matrix of numerics. The covariates of interest.}

\item{ctl}{A vector of logicals of length \code{ncol(Y)}. If
position i is \code{TRUE} then position i is considered a
negative control.}

\item{k}{A non-negative integer.The number of unobserved
confounders. If not specified and the R package sva is
installed, then this function will estimate the number of
hidden confounders using the methods of Buja and Eyuboglu
(1992).}

\item{cov_of_interest}{A vector of positive integers. The column
numbers of the covariates in X whose coefficients you are
interested in. The rest are considered nuisance parameters and
are regressed out by OLS.}

\item{include_intercept}{A logical. If \code{TRUE}, then it will
check \code{X} to see if it has an intercept term. If not, then
it will add an intercept term. If \code{FALSE}, then \code{X}
will be unchanged.}

\item{limmashrink}{A logical. Should we apply hierarchical
shrinkage to the variances (\code{TRUE}) or not (\code{FALSE})?
If \code{degrees_freedom = NULL} and \code{limmashrink = TRUE}
and \code{likelihood = "t"}, then we'll also use the limma
returned degrees of freedom.}

\item{gls}{A logical. Should we use generalized least squares
(\code{TRUE}) or ordinary least squares (\code{FALSE}) for
estimating the confounders? The OLS version is equivalent to
using RUV4 to estimate the confounders.}

\item{fa_func}{A factor analysis function. The function must have
as inputs a numeric matrix \code{Y} and a rank (numeric scalar)
\code{r}. It must output numeric matrices \code{alpha} and
\code{Z} and a numeric vector \code{sig_diag}. \code{alpha} is
the estimate of the coefficients of the unobserved confounders,
so it must be an \code{r} by \code{ncol(Y)} matrix. \code{Z}
must be an \code{r} by \code{nrow(Y)} matrix. \code{sig_diag}
is the estimate of the column-wise variances so it must be of
length \code{ncol(Y)}. The default is the function
\code{pca_naive} that just uses the first \code{r} singular
vectors as the estimate of \code{alpha}. The estimated
variances are just the column-wise mean square.}

\item{fa_args}{A list. Additional arguments you want to pass to
fa_func.}
}
\value{
\code{betahat} The estimates of the coefficients of
    interest. The values corresponding to control genes are 0.

    \code{sebetahat_unadjusted} The unadjusted standard errors of
    \code{betahat}. The values corresponding to control genes are
    \code{NA}.

    \code{tstats_unadajusted} The t-statistics corresponding to the
    coefficients of interest. These use \code{sebetahat_unadjusted}
    as the standard errors. The values corresponding to control
    genes are \code{NA}.

    \code{pvalues_unadajusted} The p-values using said statistics
    above.

    \code{sebetahat_adjusted} The unadjusted standard errors of
    \code{betahat}. This equals \code{sebetahat_unadjusted *
    multiplier}. The values corresponding to control genes are
    \code{NA}.

    \code{tstats_adjusted} The t-statistics corresponding to the
    coefficients of interest. These use \code{sebetahat_adjusted}
    as the standard errors. The values corresponding to control
    genes are \code{NA}.

    \code{pvalues_unadajusted} The p-values using said statistics
    above.

    \code{betahat_ols} The ordinary least squares (OLS) estimates
    for all of the coefficients.

    \code{sebetahat_ols} The standard errors from OLS regression.

    \code{tstats_ols} The t-statistics from OLS regression.

    \code{pvalues_ols} The p-values from OLS regression.

    \code{sigma2_unadjusted} The unadjusted variance estimates.

    \code{sigma2_adjusted} The adjusted variance estimates. This is
    equal to \code{sigma2_unadjusted * multiplier}.

    \code{Zhat} The estimates of the confounders.

    \code{alphahat} The esitmates of the coefficients of the confounders.

    \code{multiplier} The estimate of the variance inflation parameter.

    \code{mult_matrix} The unscaled covariance of \code{betahat}
    after including the confounders.

    \code{mult_matrix_ols} The OLS version of \code{mult_matrix}.

    \code{degrees_freedom} The degrees of freedom used when
    calculating the p-values.
}
\description{
This can be viewed both as a version of RUV2 and a version of RUV4.
}
\author{
David Gerard
}

