% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvimpute.R
\name{ruvem}
\alias{ruvem}
\title{Same as ruvimpute but only does ruvem and comes up with estimates of standard errors.}
\usage{
ruvem(
  Y,
  X,
  ctl,
  k = NULL,
  impute_args = list(),
  cov_of_interest = ncol(X),
  include_intercept = TRUE,
  gls = TRUE
)
}
\arguments{
\item{Y}{A matrix of numerics. These are the response variables
where each column has its own variance. In a gene expression
study, the rows are the individuals and the columns are the
genes.}

\item{X}{A matrix of numerics. The covariates of interest.}

\item{ctl}{A vector of logicals of length \code{ncol(Y)}. If
position i is \code{TRUE} then position i is considered a
negative control.}

\item{k}{The rank of the underlying matrix. Used by
\code{\link{hard_impute}} if that is the value of
\code{impute_func}. If not provided, will be estimated
by \code{\link[sva]{num.sv}}.}

\item{impute_args}{A list of additional parameters to pass to
\code{impute_func}.}

\item{cov_of_interest}{A vector of positive integers. The column
numbers of the covariates in X whose coefficients you are
interested in. The rest are considered nuisance parameters and
are regressed out by OLS.}

\item{include_intercept}{A logical. If \code{TRUE}, then it will
check \code{X} to see if it has an intercept term. If not, then
it will add an intercept term. If \code{FALSE}, then \code{X}
will be unchanged.}

\item{gls}{A logical. Should we use GLS (\code{TRUE}) or OLS
(\code{FALSE})?}
}
\description{
This is more friendly than \code{\link{ruvimpute}}.
}
\references{
\itemize{
  \item{Gerard, David, and Matthew Stephens. 2021. "Unifying and Generalizing Methods for Removing Unwanted Variation Based on Negative Controls." \emph{Statistica Sinica}, 31(3), 1145-1166. \doi{10.5705/ss.202018.0345}}
}
}
\seealso{
\code{\link{ruvimpute}}, \code{\link{em_miss}}.
}
\author{
David Gerard
}
