% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv3.R
\name{ruvimpute}
\alias{ruvimpute}
\title{General imputation framework.}
\usage{
ruvimpute(Y, X, ctl, impute_func, impute_args = list(),
  cov_of_interest = ncol(X), include_intercept = TRUE)
}
\arguments{
\item{Y}{A matrix of numerics. These are the response variables
where each column has its own variance. In a gene expression
study, the rows are the individuals and the columns are the
genes.}

\item{X}{A matrix of numerics. The covariates of interest.}

\item{ctl}{A vector of logicals of length \code{ncol(Y)}. If
position i is \code{TRUE} then position i is considered a
negative control.}

\item{impute_func}{A function that takes as inputes matrices
\code{Y21}, \code{Y22}, and \code{Y31}, and outputs \code{Y32}
and \code{sig_diag}. Additional arguments may be passed through
\code{impute_args}.}

\item{impute_args}{A list of additional parameters to pass to \code{impute_func}.}

\item{cov_of_interest}{A vector of positive integers. The column
numbers of the covariates in X whose coefficients you are
interested in. The rest are considered nuisance parameters and
are regressed out by OLS.}

\item{include_intercept}{A logical. If \code{TRUE}, then it will
check \code{X} to see if it has an intercept term. If not, then
it will add an intercept term. If \code{FALSE}, then \code{X}
will be unchanged.}
}
\description{
General imputation framework.
}
\author{
David Gerard
}

