% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_my_p.R
\name{smell_my_p}
\alias{smell_my_p}
\title{Wrapper for a Kolmogorov-Smirnov test using the p-values from the
control genes.}
\usage{
smell_my_p(pvalues, ctl)
}
\arguments{
\item{pvalues}{A vector of numbers between 0 and 1. The pvalues.}

\item{ctl}{A vector of logicals of the same length of
\code{pvalues}. A \code{TRUE} indicates that a p-value is from
a control gene and so should be uniformly distributed.}
}
\description{
Wrapper for a Kolmogorov-Smirnov test using the p-values from the
control genes.
}
\author{
David Gerard
}
