% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_my_p.R
\name{smell_my_stats}
\alias{smell_my_stats}
\title{Wrapper for a Kolmogorov-Smirnov test using the statistics from the
control genes.}
\usage{
smell_my_stats(stats, ctl, cdf_func, ...)
}
\arguments{
\item{stats}{A vector of numerics. The statistics.}

\item{ctl}{A vector of logicals of the same length of
\code{pvalues}. A \code{TRUE} indicates that a p-value is from
a control gene and so should be uniformly distributed.}

\item{cdf_func}{A cdf function.}

\item{...}{More parameters to pass to \code{qfunc}.}
}
\description{
Wrapper for a Kolmogorov-Smirnov test using the statistics from the
control genes.
}
\author{
David Gerard
}
