% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv3.R
\name{softimpute_wrapper}
\alias{softimpute_wrapper}
\title{A wrapper for using the softImpute function from the softImpute package.}
\usage{
softimpute_wrapper(Y)
}
\arguments{
\item{Y}{A matrix with missing values.}
}
\value{
A matrix with the missing values imputed.
}
\description{
A wrapper for using the softImpute function from the softImpute package.
}
\author{
David Gerard
}

