% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv_ash.R
\name{tregress_em}
\alias{tregress_em}
\title{EM algorithm to find regression coefficients using t-likelihood
when variances are known up to scale.}
\usage{
tregress_em(Y, alpha, sig_diag, nu, lambda_init = NULL, Z_init = NULL,
  control_args = list())
}
\arguments{
\item{Y}{A matrix of numerics with one column. The response
variables.}

\item{alpha}{A matrix of numerics, the covariates. It must be that
\code{nrow(Y)} is equal to \code{nrow(alpha)}.}

\item{sig_diag}{A vector of numerics. The variances of the elements
in \code{Y}, but only assumed to be known up to a scaling
factor.}

\item{nu}{A positive numeric scalar. The known degrees of freedom
of the t-distribution.}

\item{lambda_init}{A positive numeric scalar. The initial value of
the variance inflation parameter. Defaults to the estimate
under Gaussian errors.}

\item{Z_init}{A matrix of numerics with one column. The initial
value of the coefficients of \code{alpha}. Defaults to the
estimate under Gaussian errors.}

\item{control_args}{A list of control arguments for the EM
algorithm that is passed to SQUAREM.}
}
\description{
When using a non-standard t-distribution for your regression
likelihood, there is a simple latent variable representation that
allows us to develop an EM algorithm. This is a very similar
procedure to Lange et al (1989).
}
\author{
David Gerard
}
\references{
Lange, Kenneth L., Roderick JA Little, and Jeremy MG
    Taylor. "Robust statistical modeling using the t distribution."
    Journal of the American Statistical Association 84.408 (1989):
    881-896.
}
\seealso{
\code{\link{tregress_obj}} for the objective function that
    this function maximizes, \code{\link{tregress_fix}} for the
    fixed point iteration that this function calls.
}

