% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv_ash.R
\name{tregress_obj_wrapper}
\alias{tregress_obj_wrapper}
\title{Wrapper for \code{\link{tregress_obj}}.}
\usage{
tregress_obj_wrapper(lambda, Z, Y, alpha, sig_diag, nu)
}
\arguments{
\item{lambda}{A positive numeric. The variance inflation parameter.}

\item{Z}{A matrix of numerics with one column. The coefficients.}

\item{Y}{A matrix of numerics with one column. The response
variables.}

\item{alpha}{A matrix of numerics, the covariates. It must be that
\code{nrow(Y)} is equal to \code{nrow(alpha)}.}

\item{sig_diag}{A vector of numerics. The variances of the elements
in \code{Y}, but only assumed to be known up to a scaling
factor.}

\item{nu}{A positive numeric scalar. The known degrees of freedom
of the t-distribution.}
}
\description{
This is mostly so I can run \code{stats::optim} with just
\code{lambda}.
}
\author{
David Gerard
}

