% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash_uniform_em.R
\name{unif_int_grad}
\alias{unif_int_grad}
\title{Gradient wrt z2 of intermediate function in EM.}
\usage{
unif_int_grad(z2, xi, betahat_ols, alpha_tilde, S_diag, a_seq, b_seq,
  qvals, degrees_freedom)
}
\arguments{
\item{z2}{A vector of numerics. The current values of the
unobserved confounders.}

\item{xi}{A positive numeric. The current value of the variance
inflation factor.}

\item{betahat_ols}{A vector of numerics. The OLS regression
coefficients.}

\item{alpha_tilde}{A matrix of numerics. The estimats of the
coefficients of the confounders.}

\item{S_diag}{A vector of positive numerics. The standard errors.}

\item{a_seq}{A vector of numerics (usually non-positive). The lower
bounds of the uniform mixing densities.}

\item{b_seq}{A vector of numerics (usually non-negative). The upper
bounds of the uniform mixing densities.}

\item{qvals}{A matrix of non-negative numerics whose rows sum to
one. The E-step proportions.}

\item{degrees_freedom}{A positive scalar or a vector of positive
scalars the length of \code{betahat_ols}. The degrees of
freedom of the t-distribution.}
}
\description{
Gradient wrt z2 of intermediate function in EM.
}
\author{
David Gerard
}
