% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash_uniform_em.R
\name{uniform_mix_fix_wrapper}
\alias{uniform_mix_fix_wrapper}
\title{Wrapper for \code{\link{uniform_mix_fix}}, mostly so I can use SQUAREM.}
\usage{
uniform_mix_fix_wrapper(pizxi_vec, betahat_ols, S_diag, alpha_tilde, a_seq,
  b_seq, lambda_seq, degrees_freedom, scale_var = TRUE)
}
\arguments{
\item{pizxi_vec}{A vector of numerics. The first M of which are
\code{pi_vals}, the next k of which are \code{z2}, the last
element is \code{xi}.}

\item{betahat_ols}{A vector of numerics. The OLS regression
coefficients.}

\item{S_diag}{A vector of positive numerics. The standard errors.}

\item{alpha_tilde}{A matrix of numerics. The estimates of the
coefficients of the confounders.}

\item{a_seq}{A vector of numerics (usually non-positive). The lower
bounds of the uniform mixing densities.}

\item{b_seq}{A vector of numerics (usually non-negative). The upper
bounds of the uniform mixing densities.}

\item{lambda_seq}{A vector of numerics greater than 1. The
penalties on the mixing proportions.}

\item{degrees_freedom}{A positive scalar or a vector of positive
scalars the length of \code{betahat_ols}. The degrees of
freedom of the t-distribution.}

\item{scale_var}{A logical. Should we update the variance inflation
parameter (\code{TRUE}) or not (\code{FALSE})}
}
\description{
Wrapper for \code{\link{uniform_mix_fix}}, mostly so I can use SQUAREM.
}
\author{
David Gerard
}
