% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{update_f}
\alias{update_f}
\title{Utility function for updating F in \code{gdfa}.}
\usage{
update_f(LY, xi, theta, ltl)
}
\arguments{
\item{LY}{A vector. A column of \code{t(L_current) \%*\% Y_current}}

\item{xi}{A numeric scalar. An element of xi_current.}

\item{theta}{A numeric scalar. An element of theta_current.}

\item{ltl}{t(L_current) \%*\% L_current}
}
\description{
Utility function for updating F in \code{gdfa}.
}
\seealso{
\code{\link{gdfa}}
}
