% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv2_fa.R
\name{update_lambda}
\alias{update_lambda}
\title{Wrapper for \code{\link[stats]{optim}} and \code{\link{qmle_obj}}.}
\usage{
update_lambda(alpha, sig_diag, lambda, Y1, Y2, maxit = 10)
}
\arguments{
\item{alpha}{The current factor loadings}

\item{sig_diag}{A vector of the current variances}

\item{lambda}{The current inflation parameter for Y1}

\item{Y1}{The first part of the observations.}

\item{Y2}{The second part of the observations.}

\item{maxit}{The number of iterations to run. Shouldn't be too
large since this is only an intermediate step.}
}
\description{
Uses Brent's method to obtimize the variance inflation parameter
given sig_diag and alpha.
}
\author{
David Gerard
}
