% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv2_fa.R
\name{update_sig_alpha_basic}
\alias{update_sig_alpha_basic}
\title{Basic no optimized code update for sig_diag and alpha.}
\usage{
update_sig_alpha_basic(alpha, sig_diag, lambda, Y1, Y2)
}
\arguments{
\item{alpha}{The current factor loadings}

\item{sig_diag}{A vector of the current variances}

\item{lambda}{The current inflation parameter for Y1}

\item{Y1}{The first part of the observations.}

\item{Y2}{The second part of the observations.}
}
\description{
Basic no optimized code update for sig_diag and alpha.
}
\author{
David Gerard
}
