% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapp.R
\name{vicarius_leapp}
\alias{vicarius_leapp}
\title{Calibrated LEAPP where the tuning parameter is chosen by a modified
cross-validation.}
\usage{
vicarius_leapp(Y, X, k = NULL, cov_of_interest = ncol(X),
  limmashrink = FALSE, include_intercept = TRUE, fa_func = pca_naive,
  fa_args = list())
}
\description{
Calibrated LEAPP where the tuning parameter is chosen by a modified
cross-validation.
}
\author{
David Gerard
}

