% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv4.R
\name{vruvinv}
\alias{vruvinv}
\title{Variance inflated RUV-inverse.}
\usage{
vruvinv(Y, X, ctl, cov_of_interest = ncol(X), likelihood = c("t", "normal"),
  include_intercept = TRUE, fa_func = pca_naive, fa_args = list(),
  adjust_bias = FALSE)
}
\arguments{
\item{Y}{A matrix of numerics. These are the response variables
where each column has its own variance. In a gene expression
study, the rows are the individuals and the columns are the
genes.}

\item{X}{A matrix of numerics. The covariates of interest.}

\item{ctl}{A vector of logicals of length \code{ncol(Y)}. If
position i is \code{TRUE} then position i is considered a
negative control.}

\item{cov_of_interest}{A vector of positive integers. The column
numbers of the covariates in X whose coefficients you are
interested in. The rest are considered nuisance parameters and
are regressed out by OLS.}

\item{likelihood}{Either \code{"normal"} or \code{"t"}. If
\code{likelihood = "t"}, then the user may provide the degrees
of freedom via \code{degrees_freedom}.}

\item{include_intercept}{A logical. If \code{TRUE}, then it will
check \code{X} to see if it has an intercept term. If not, then
it will add an intercept term. If \code{FALSE}, then \code{X}
will be unchanged.}

\item{fa_func}{A factor analysis function. The function must have
as inputs a numeric matrix \code{Y} and a rank (numeric scalar)
\code{r}. It must output numeric matrices \code{alpha} and
\code{Z} and a numeric vector \code{sig_diag}. \code{alpha} is
the estimate of the coefficients of the unobserved confounders,
so it must be an \code{r} by \code{ncol(Y)} matrix. \code{Z}
must be an \code{r} by \code{nrow(Y)} matrix. \code{sig_diag}
is the estimate of the column-wise variances so it must be of
length \code{ncol(Y)}. The default is the function
\code{pca_naive} that just uses the first \code{r} singular
vectors as the estimate of \code{alpha}. The estimated
variances are just the column-wise mean square.}

\item{fa_args}{A list. Additional arguments you want to pass to
fa_func.}

\item{adjust_bias}{A logical. Should we also use the control genes
to adjust for bias (\code{TRUE}) or not (\code{FALSE})}
}
\value{
See \code{\link{vruv4}} for the elements returned.
}
\description{
RUV-inverse as described in Gagnon-Bartsch et al (2013) is just
RUV4 using the maximum number of confounders allowed by the model,
followed by estimating the variances using a method-of-moments type
approach. \code{vruvinv} is similar in spirit to this by using the
maximum number of confounders allowed by the model, but we still
estimate the variance inflation using the confounders. We force
\code{limmashrink = TRUE}, otherwise there is only one degree of
freedom to estimate the variance.
}
\details{
Hence, this is just a wrapper for \code{\link{vruv4}}, but with a
special choice for the number of confounders and forcing
\code{limmashrink = TRUE}.
}
\author{
David Gerard
}
\references{
Gagnon-Bartsch, J., Laurent Jacob, and Terence
    P. Speed. "Removing unwanted variation from high dimensional
    data with negative controls."
    Berkeley: Department of Statistics. University of California
    (2013).
}

