% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvb.R
\name{bfl}
\alias{bfl}
\title{Simple Bayesian low rank matrix decomposition.}
\usage{
bfl(
  Y21,
  Y31,
  Y32,
  k,
  nsamp = 10000,
  burnin = round(nsamp/4),
  keep = 20,
  print_update = TRUE,
  plot_update = FALSE,
  rho_0 = 0.1,
  alpha_0 = 0.1,
  beta_0 = 1,
  eta_0 = 0.1,
  tau_0 = 1
)
}
\arguments{
\item{Y21}{Top left of matrix.}

\item{Y31}{Bottom left of matrix.}

\item{Y32}{Top right of matrix.}

\item{k}{The rank of the mean matrix.}

\item{nsamp}{A positive integer. The number of samples to draw.}

\item{burnin}{A positive integer. The number of early samples to
discard.}

\item{keep}{A positive integer. We will same the updates of
\code{Y22} every \code{keep} iteration of the Gibbs sampler.}

\item{print_update}{A logical. Should we print a text progress bar
to keep track of the Gibbs sampler (\code{TRUE}) or not
(\code{FALSE})?}

\item{plot_update}{A logical. Should we make some plots to keep
track of the Gibbs sampler (\code{TRUE}) or not (\code{FALSE})?}

\item{rho_0}{A scalar. The prior "sample size" for the precisions.}

\item{alpha_0}{A scalar. The prior "sample size" for the mean of the precisions.}

\item{beta_0}{A scalar. The prior mean of the precisions.}

\item{eta_0}{A scalar. The prior "sample size" for the scale of the mean matrix.}

\item{tau_0}{A scalar. The prior mean of the scale of the mean matrix.}
}
\description{
"bfl" = "Bayesian factor loading"
}
\details{
This is as simple as they come. I put normal priors on the loadings
and factors and gamma priors on the precisions. The hyperparameters
are set to provide weak prior information by default.

The main difference between this version and others is that the
factors are a prior assumed to have the same variances as the data
observations. This might be distasteful to some.

There is no parameter expansion in this one. To see one with
parameter expansion, and a much faster version, see
\code{\link{bfa_gs_linked}}.
}
\references{


\itemize{
  \item{Gerard, David, and Matthew Stephens. 2021. "Unifying and Generalizing Methods for Removing Unwanted Variation Based on Negative Controls." \emph{Statistica Sinica}, 31(3), 1-22. \doi{10.5705/ss.202018.0345}}
}
}
\seealso{
\code{\link{bfa_gs_linked}}.
}
\author{
David Gerard
}
