% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash_normal_em.R
\name{brent_obj_norm}
\alias{brent_obj_norm}
\title{Objective function for updating variance inflation parameter during
EM for mixtures of normals prior.}
\usage{
brent_obj_norm(xi, qvals, S_diag, tau2_seq, resid_vec)
}
\arguments{
\item{xi}{The current value of the variance inflation parameter.}

\item{qvals}{The probabilities from the E step. A matrix.}

\item{S_diag}{A vector of positive numerics. The estimated standard
errors.}

\item{tau2_seq}{The grid of variances. This is the same thing as
grid_seq in \code{\link{mouthwash_second_step}} if
\code{likelihood = "normal"}.}

\item{resid_vec}{The residuals from the OLS estimates and the bias
term induced by the confounders.}
}
\description{
Objective function for updating variance inflation parameter during
EM for mixtures of normals prior.
}
\author{
David Gerard
}

