% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv4.R
\name{cruv4}
\alias{cruv4}
\title{RUV4's second step.}
\usage{
cruv4(betahat_ols, alpha_scaled, sig_diag_scaled, ctl, degrees_freedom,
  gls = TRUE, likelihood = "t")
}
\arguments{
\item{betahat_ols}{A matrix of numerics with one column. The ols
estimates of beta.}

\item{alpha_scaled}{A matrix of numerics. The estimated
coefficients of the unobserved confounders.}

\item{sig_diag_scaled}{A vector of positive numerics. The estimated
standard errors of \code{betahat_ols}.}

\item{ctl}{A vector of logicals of length \code{ncol(Y)}. If
position i is \code{TRUE} then position i is considered a
negative control.}

\item{degrees_freedom}{A positive numeric. The degrees of freedom
of the t-likelihood if using it.}

\item{gls}{A logical. Should we use generalized least squares
(\code{TRUE}) or ordinary least squares (\code{FALSE}) for
estimating the confounders? The OLS version is equivalent to
using RUV4 to estimate the confounders.}

\item{likelihood}{Either \code{"normal"} or \code{"t"}. If
\code{likelihood = "t"}, then the user may provide the degrees
of freedom via \code{degrees_freedom}.}
}
\description{
This is the old version that only works for one covariate of
interest. Use \code{\link{cruv4_multicov}} for the more recent
version.
}
\author{
David Gerard
}
