% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_miss.R
\name{em_miss}
\alias{em_miss}
\title{EM algorithm for factor analysis with missing block matrix.}
\usage{
em_miss(Y21, Y31, Y32, k, gls = TRUE)
}
\arguments{
\item{Y21}{Top left of matrix.}

\item{Y31}{Bottom left of matrix.}

\item{Y32}{Top right of matrix.}

\item{k}{The rank of the mean matrix.}

\item{gls}{A logical. Should we estimate Z by generalized least
squares (\code{TRUE}) or by a multivariate normality assumption
(\code{FALSE})?}
}
\value{
The top right of the matrix.
}
\description{
EM algorithm for factor analysis with missing block matrix.
}
\author{
David Gerard
}

