% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvimpute.R
\name{knn_wrapper}
\alias{knn_wrapper}
\title{Wrapper for impute.knn}
\usage{
knn_wrapper(Y)
}
\arguments{
\item{Y}{A matirx with missing values.}
}
\value{
A matrix with the missing values imputed.
}
\description{
Wrapper for impute.knn
}
\seealso{
\code{\link[impute]{impute.knn}}
}
\author{
David Gerard
}
