% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash.R
\name{mouthwash_second_step}
\alias{mouthwash_second_step}
\title{The second step of MOUTHWASH.}
\usage{
mouthwash_second_step(betahat_ols, S_diag, alpha_tilde, grid_seq, lambda_seq,
  mixing_dist = c("uniform", "+uniform", "sym_uniform", "normal"),
  likelihood = c("t", "normal"), pi_init_type = c("zero_conc", "uniform",
  "random"), scale_var = TRUE)
}
\arguments{
\item{betahat_ols}{A vector of numerics. The OLS estimates of the
coefficients of interest.}

\item{S_diag}{A vector of positive numerics. The estimated standard
errors.}

\item{alpha_tilde}{A matrix. The number of rows should be equal the
length of betahat_ols. The number of columns should equal the
number of hidden confounders.}

\item{grid_seq}{The grid of variances (if \code{mixing_dist =
"normal"}) or the grid of uniform bounds. See
\code{\link{mouthwash}} for details.}

\item{lambda_seq}{A numeric vector with elements all greater than
or equal to 1. These are the tuning parameters for the mixing
proportions. This can only be specified if \code{grid_seq} is
also specified.}

\item{mixing_dist}{A character. Should we use a mixture of uniforms
(\code{"uniform"}), a mixture of uniforms with minimum at 0
(\code{"+uniform"}), a mixture of uniforms symmetric at 0
(\code{"sym_uniform"}), or a mixture of normals
(\code{"normal"})?}

\item{likelihood}{Either \code{"normal"} or \code{"t"}. If
\code{likelihood = "t"}, then the user may provide the degrees
of freedom via \code{degrees_freedom}.}

\item{pi_init_type}{How should we initialize the mixing
proportions? By concentrating on zero (\code{"zero_conc"}), by
equal weights on all mixing distributions (\code{"uniform"}),
or by sampling uniformly on the simplex (\code{"random"})?}

\item{scale_var}{A logical. Should we estimate a variance inflation
parameter (\code{TRUE}) or not (\code{FALSE})?}
}
\description{
The second step of MOUTHWASH.
}
\author{
David Gerard
}

