% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouthwash_uniform_em.R
\name{ptdiff_mat}
\alias{ptdiff_mat}
\title{More stable way to calculate differences in T cdfs when input is a matrix.}
\usage{
ptdiff_mat(X, Y, degrees_freedom)
}
\arguments{
\item{X}{A matirx of quantiles of the t.}

\item{Y}{Another matirx of quantiles of the t.}

\item{degrees_freedom}{The degrees of freedom of the t.}
}
\description{
More stable way to calculate differences in T cdfs when input is a matrix.
}
\seealso{
\code{\link{ptdiff}}.
}
\author{
David Gerard
}
