% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv3.R
\name{ruvimpute}
\alias{ruvimpute}
\title{General imputation framework.}
\usage{
ruvimpute(Y, X, ctl, impute_func = missforest_wrapper, impute_args = list(),
  cov_of_interest = ncol(X), include_intercept = TRUE,
  do_variance = FALSE)
}
\arguments{
\item{Y}{A matrix of numerics. These are the response variables
where each column has its own variance. In a gene expression
study, the rows are the individuals and the columns are the
genes.}

\item{X}{A matrix of numerics. The covariates of interest.}

\item{ctl}{A vector of logicals of length \code{ncol(Y)}. If
position i is \code{TRUE} then position i is considered a
negative control.}

\item{impute_func}{A function that takes as input a matrix names
\code{Y} that has missing values and returns a matrix called
\code{Yhat} of the same dimension of \code{Y} with the missing
values filled in. If \code{do_variance = TRUE}, then
\code{impute_func} should also return \code{sig_diag} --- a
vector of column-specific variance estimates. The default is a
wrapper for \code{\link[softImpute]{softImpute}}.}

\item{impute_args}{A list of additional parameters to pass to
\code{impute_func}.}

\item{cov_of_interest}{A vector of positive integers. The column
numbers of the covariates in X whose coefficients you are
interested in. The rest are considered nuisance parameters and
are regressed out by OLS.}

\item{include_intercept}{A logical. If \code{TRUE}, then it will
check \code{X} to see if it has an intercept term. If not, then
it will add an intercept term. If \code{FALSE}, then \code{X}
will be unchanged.}

\item{do_variance}{A logical. Does \code{impute_func} also return
estimates of the column-specific variances?}
}
\value{
\code{beta2hat} The estimates of the coefficients of the
    covariates of interest that do not correspond to control genes.

    \code{betahat_long} The estimates of the coefficients. Those
    corresponding to control genes are set to 0.

    \code{sebetahat} If \code{do_variance = TRUE}, then these are
    the "standard errors" of \code{beta2hat} (but not really).

    \code{tstats} If \code{do_variance = TRUE}, then these are
    the "t-statistics" of \code{beta2hat} (but not really).

    \code{pvalues} If \code{do_variance = TRUE}, then these are
    the "p-values" of \code{tstats} (but not really).
}
\description{
General imputation framework.
}
\author{
David Gerard
}

