% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruvimpute.R
\name{softimpute_wrapper}
\alias{softimpute_wrapper}
\title{A wrapper for using the softImpute function from the softImpute package.}
\usage{
softimpute_wrapper(Y, max_rank)
}
\arguments{
\item{Y}{A matrix with missing values.}

\item{max_rank}{The same as the \code{max.rank} option in
\code{\link[softImpute]{softImpute}}.}
}
\value{
A matrix with the missing values imputed.
}
\description{
A wrapper for using the softImpute function from the softImpute package.
}
\seealso{
\code{\link[softImpute]{softImpute}}
}
\author{
David Gerard
}
