% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapp.R
\name{vicarius_principis_IPOD}
\alias{vicarius_principis_IPOD}
\title{Modified version of leapp::IPOD to run a version of
cross-validation to choose the tuning parameter.}
\usage{
vicarius_principis_IPOD(X, Y, H, sigma, method = "hard", TOL = 1e-04)
}
\description{
Modified version of leapp::IPOD to run a version of
cross-validation to choose the tuning parameter.
}

